# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/channel/create": {
            "post": {
                "description": "创建一个新的渠道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "创建渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Channel",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    },
                    "400": {
                        "description": "无效的输入或验证错误",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/channel/delete/{id}": {
            "delete": {
                "description": "根据提供的ID删除现有的渠道",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "删除渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/channel/get/{id}": {
            "get": {
                "description": "根据提供的ID获取单个渠道的详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "根据ID获取渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    },
                    "404": {
                        "description": "渠道未找到",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/channel/list": {
            "get": {
                "description": "获取所有渠道的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "获取所有渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Channel"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/channel/test": {
            "post": {
                "description": "测试推送",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "测试推送",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Channel",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "无效的输入或验证错误",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/channel/update/{id}": {
            "put": {
                "description": "根据提供的ID更新现有的渠道",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channel"
                ],
                "summary": "更新渠道",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel",
                        "name": "channel",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    },
                    "400": {
                        "description": "无效的输入或验证错误",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Channel": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "渠道配置（统一结构）",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "description": "UUID",
                    "type": "integer"
                },
                "name": {
                    "description": "渠道名称（如：Email-SMTP、OneBot-v11）",
                    "type": "string"
                },
                "type": {
                    "description": "渠道类型",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ChannelType"
                        }
                    ]
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "types.ChannelType": {
            "type": "string",
            "enum": [
                "Bark",
                "DingTalk",
                "FeiShu",
                "OneBot",
                "Gotify",
                "ServerChan",
                "PushDeer",
                "Synology",
                "IYUU",
                "PushPlus",
                "QMSG",
                "WeCom",
                "Telegram",
                "Email",
                "Webhook",
                "Ntfy",
                "WxPusher"
            ],
            "x-enum-comments": {
                "TypeBark": "Bark",
                "TypeDingTalk": "钉钉机器人",
                "TypeEmail": "邮箱",
                "TypeFeiShu": "飞书机器人",
                "TypeGotify": "Gotify",
                "TypeIYUU": "IYUU",
                "TypeNtfy": "Ntfy",
                "TypeOneBot": "OneBot",
                "TypePushDeer": "PushDeer",
                "TypePushPlus": "PushPlus",
                "TypeQMSG": "Qmsg酱",
                "TypeServerChan": "Server酱",
                "TypeSynology": "群晖chat",
                "TypeTelegram": "Telegram",
                "TypeWeCom": "企业微信",
                "TypeWebhook": "Webhook",
                "TypeWxPusher": "WxPusher"
            },
            "x-enum-varnames": [
                "TypeBark",
                "TypeDingTalk",
                "TypeFeiShu",
                "TypeOneBot",
                "TypeGotify",
                "TypeServerChan",
                "TypePushDeer",
                "TypeSynology",
                "TypeIYUU",
                "TypePushPlus",
                "TypeQMSG",
                "TypeWeCom",
                "TypeTelegram",
                "TypeEmail",
                "TypeWebhook",
                "TypeNtfy",
                "TypeWxPusher"
            ]
        },
...
    }
}